Removing the features list and iterator
The features list and iterator are basically re-implementations of slices. In order to make the code base smaller, I removed it and instead implemented their functionality with slices.

However, this involves the allocation of a vector in the `instantiate` function, but this is okay since `instantiate` is only called once for a plugin instance and not within a real-time-sensitive context.