Rearranging the feature-related modules
This involved moving the feature descriptor and demoting it from a proper feature representation to a convenient intermediate type for a client's feature detection phase.

Both on the host's side and the client's side, the proper representation of a feature should be a reference to the proper struct and the "transport" representation of a feature should be the `repr(C)` descriptor. An intermediate representation has only very few use cases.