Implementing a new feature description system
I generally like the idea how host features are provided by the user: Provide a struct that is constructed from the features list and contains the pointers to all required features.
Creating this struct manually would require quite some boilerplate code and is therefore a candidate for a macro.

However, it would be hard to create a macro that could distinguish required features from optional ones. I tried to find a way to support optional features using the custom-struct-based approach and did not find one.
Therefore, I introduced a different approach: The plugin instance creates a map-based wrapper which gives access to the features in a safe way. Using this wrapper, one could create a macro to collect required features in a struct, while still having the option to manually find and save required or optional features.

# Safety

Technically, there is a safe method that does something unsafe, but `FeatureContainer` can only be created by the plugin instance. Therefore, it can only contain pointers supplied by the host. Since we've also assured that the pointers are not null, we can not safe ourselves from invalid pointers supplied by the host and may therefore assume that they are valid. Therefore, this code is sound, according to the usual assumptions.