Removing the URI type
The URI type does not do something special, it is only a CStr with another name. It uses a lot of code only to recreate the features of a CStr. The only reason why this type may be useful is that it should be able to identify valid and invalid URIs. However, this is a task for a different crate altogether and therefore should not be done in this one. Also, we don't really need it: To us, a URI is only a string to identify certain things, we don't really need to care if it's a valid URI a web browser understands or not.

The real cause of the removal is another problem: In order to be useful, the URI needs to implement both `Hash` and `PartialEq`. Clippy does not like that. On the other hand, `CStr` implements both traits and therefore, it is easier to simply use a CStr.