Add patch* uris
Issue: LV2 has for its URIs the convention that types are CamelCased and data
are mixedCased. In rust-lv2 URIs appear as structs (CamelCases) as well as
fields (snake_cased) of an URIDCollection struct.

This leads to the clash between the LV2_PATCH__Request and LV2_PATCH__request
uridd. Further more there is the LV2_PATCH__Move URI that leads to the field
name `move` clashing` with Rusts keyword `move`.

Suggestion: Follow LV2's convention and #![allow(non_camel_case_types, non_snake_case)]
in */uris.rs

We need to agree if we follow the LV2 convention consequently and use mixedCase
for URI data or if we stick to Rust's convention and use snake_case.